Fixed collision events in xsge.gui to only happen with one wall.
Before, it was doing collision events with whatever walls happened
to be collided with on checking. It now only does collision events
on the very *last* one.

This has a side effect of guaranteeing a collision with only one wall,
but that's a lot better than 1. what walls are being collided with
having no consistency, and 2. colliding with walls that are *behind*
the wall that you're actually colliding with.