Fixed keyboard focused window not being completely destroyed.
The only code which was getting rid of the reference depended on another
window existing. This caused the last window to be destroyed to have a
sort of phantom presence: it wasn't visible, but it was still accepting
keyboard input and having whatever effects it would have if it were
visible. This has been fixed.