Made xsge_path.Path precise.
More specifically, when an object reaches one of the points, its
position is set to precisely that point. This prevents little differences
caused by delta timing which are very small, but can add up if
several paths are followed in a row, or if a path is looped over
several times.