Removed an inefficient practice in rendering static tiles.
Previously, they were just made into objects. Now, if possible, they
are instead drawn to a room-sized sprite which is then given to one
object. This turned out to be an exceptional speed boost in reTux,
especially level 2, which is quite heavy on tiles.

Incidentally, I seem to have come across some sort of bug in the
SGE while doing this: destroying objects in the create event of the
new RenderedTiles class was preventing the create event of an
unrelated object from executing, for some reason. This needs to be
investigated.