Did something to fix the semi-infinite loop sometimes caused by slopes.
Investigation showed that the infinite loops were caused by managing
to push the object forward by infintesimal amounts (for example, in
one case I recorded increments of 1/300,000,000,000,000 of a pixel),
and this was preventing the slope that actually mattered at this point
from being given its turn. I'm thinking this was caused by imprecise
floating point arithmetic. I've attempted to fix this by rounding the
target end location of a given slope. Time will tell if this actually
fixed the bug.

There are still more bugs to fix in this new slope implementation:
firstly, there's a strange bug that causes the character in a
platformer to fall through the floor after the slope a couple frames
after leaving the slope; secondly, there's a bug that causes the
character to teleport from the bottom-right corner to the top-left
corner of a top-right slope; and thirdly, there's a bug that causes
the character to not stick properly to a pair of slopes when moving
down from a less steep one to a more steep one.