Improved the way darkness projection works.
Basically, I just changed it to use window projections instead of
room projections. This meant the removal of the "z" argument, of
course.

The main reason I did this was because I was having trouble keeping
the darkness aligned with the view in ReTux, and that looked quite
ugly. With this change, the worst games could have to deal with is
lights being slightly offset from their light sources, which
shouldn't look anywhere near as ugly.