Fixed a typo in xsge_physics ("sope" instead of "slope").
This is a start at an attempt to fix the existing, basically working
slope physics, rather than rewrite them. I've been at the rewrite
(now under the physics-rewrite branch) for two months now, and the
damn thing is *still* more buggy than what I had before, with
virtually no progress made. What's more, I don't even know if any
of the bugs I set out to fix actually were fixed, with one exception.
The rewrite is just so complex, it's incredibly difficult to debug.

So in light of that, I'm going to attempt to fix the four major
problems in these slope physics:

- Collision events of the other axis being called (which, in ReTux,
has resulted in bottom slopes stopping objects that were moving up,
because ReTux uses this event to set yvelocity to 0). This one
should be easy to solve.

- Strange slope configurations, like the one in ReTux's first test
level, not stopping movement because the other slope is used to make
the object pass through.

- Switching from a gentle slope to a steep slope causing loss of
floor contact. For this one, I'll want to take a look at what I did
right in the rewrite, and incorporate it here -- or I could just take
a simpler approach to ensure that the object sticks to *something* if
the slope being moved along is "sticky".

- Switching from one slope to another of the same type not working, for some
reason. I suppose it probably has to do with the secondary effect
slopes have where they stop movement if you touch their corners.
Though considering some spikies in one level of ReTux don't turn around
when this happens to them, it might be somewhat more complicated.