Fixed failure to properly stick to slopes when moving down.
The cause of this particular bug was so simple, it's hilarious: the
functions getting x and y positions of slopes was simplistic, making
a (false) assumption that the slopes were lines of infinite length.
The intent was that you were supposed to check for this yourself,
but the code that was supposed to be sticking you to stuff wasn't
doing that, and that caused it to choose the wrong slope and actually
fail to push you down as far as you could be pushed down. Fixed.

Interestingly, it seems that the other major bug -- the one where
your movement was stopped when moving up -- has been exacerbated by
fixing this bug. So the imperfect code in question was actually
making that bug substantially less prevalent.