Solved one cause of the bug stopping you from moving up slopes.
Yep, it seems there are multiple causes. This wone was quite simple:
downward movement because of gravity was moving you slightly down,
effectively canceling some movement. Not a problem, except that
when reaching the top of a slope, you stopped there, so there was
a constant battle between reaching the top of the slope and moving
a little bit down. This aspect has been fixed by adding code that
causes movement to continue when you reach the top of the slope.

Now, for the other cause...