Fixed the other half of the bug that got you stuck at the top of slopes.
It's working!!!!! I am so ecstatic right now!

This, too, was a really simple thing, it turned out: slope behind
was being checked first, and that meant that it was blocking
movement on account of being a corner collision. This has now been
accounted for by making sure the slope which is furthest in the
opposite direction of the one you're moving in is always chosen
first.