Fixed a bizarre slight text offset bug in xsge_gui.
I have no idea what I was thinking with the equation that caused
this. It doesn't make any sense whatsoever; "ih" in that context
is supposed to be the height of the whole space that items can
occupy, but this equation, as far as I can tell, did absolutely
nothing except cause the amount of vertical space available to
be very slightly overestimated; the fewer menu items, the less
difference it made (with only one menu item, it made no difference).

Removing this line doesn't seem to cause any problems, so that's
what I have done.