Changed the priority of the way class is determined.
A long-standing quirk in ReTux has been that putting wrong objects
in the "solid" layer would cause the game to crash because the
"solid" class didn't have keyword arguments being specified. In
hindsight, it makes no sense for the layer's name to take priority
over what is specified for that particular tile, so they have swapped
places to resolve this and other similar problems.