Added slope-induced acceleration.
This is designed to better support a "hills" effect, where going up
is hard and going down is easy, except this focuses on the acceleration
aspect. It allows for an effect where hills can be used to pick up
speed faster or force you to backtrack and get a running start.