Initial version of xsge_tiled.
Someone pointed out to me that Tiled has a JSON format now:

https://doc.mapeditor.org/en/stable/reference/json-map-format/

So this new xSGE module will implement usage of that, based on the
old xsge_tmx code and done in a compatible way. For now it's just
a copy of xsge_tmx, but I'll be removing its usage of the tmx
library and replacing that with json calls.

I'm really, really glad we found out about this.

🕵️