Relicensed all of xSGE to the GNU LGPL.
This is possible since we (The Diligent Circle) are the only author
of copyrightable code in xSGE. It's something we've been thinking of
doing since leaving that nasty FSF cult. There's no good reason to
make xSGE only available to open source games. Even the LGPL's
restriction on static compiling is overly restrictive.